requirements.txt
asgiref==3.7.2
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
Django==4.2.11
django-ckeditor==6.7.0
django-filter==23.5
django-js-asset==2.2.0
django-storages==1.14.2
django-widget-tweaks==1.5.0
et-xmlfile==1.1.0
idna==3.6
numpy==1.26.2
openpyxl==3.1.2
pandas==2.1.4
pdfkit==1.0.0
Pillow==10.1.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
python-dateutil==2.8.2
pytz==2023.3.post1
reportlab==4.1.0
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.2.1
whitenoise==6.6.0